fix(ci): resolve all mypy type errors and CI/CD pipeline issues
- Fix Optional type annotations for function parameters
- Resolve device attribute access issues in Triton kernels
- Fix MPS memory estimation with proper fallback methods
- Add null checks for CUDA stream synchronization
- Replace bare except with specific Exception handling
- Fix type conversion issues for float/int assignments
- Update torch import handling in test modules

All CI/CD checks now pass.

Pipeline ready for production deployment.

+++ src/specdec/pipeline.py
@@ -992,11 +992,15 @@ class SpeculativePipeline(SpeculativeDecoder):
             mem_rss_mb = process.memory_info().rss / 1024 / 1024
 
             # Get CUDA memory if available
-            cuda_mem_allocated = 0
-            cuda_mem_peak = 0
+            cuda_mem_allocated = 0.0
+            cuda_mem_peak = 0.0
             if self.device == "cuda" and torch.cuda.is_available():
-                cuda_mem_allocated = torch.cuda.memory_allocated() / 1024 / 1024  # MB
-                cuda_mem_peak = torch.cuda.max_memory_allocated() / 1024 / 1024  # MB
+                cuda_mem_allocated = float(
+                    torch.cuda.memory_allocated() / 1024 / 1024
+                )  # MB
+                cuda_mem_peak = float(
+                    torch.cuda.max_memory_allocated() / 1024 / 1024
+                )  # MB
 
             # Get actual dtype from optimization manager if available
             actual_dtype = str(self.dtype).replace("torch.", "")
